LPS-55504 move to module
