+++ modules/portal/portal-scripting-api/src/com/liferay/portal/scripting/ClassVisibilityChecker.java
